Several changes
Changed random_double to random_real for adapting precision when compiling.
rgb's as spectrum work. sampled_spectrum takes a lot of time.
It needs to be able to work with smaller samples.
Besides that pdf code should be added to ray color function to sample
rays towards light and model.