Strategy for spectral rendering
Each ray comes with a random wavelength.
And at the end of ray color function we obtain a power response.
This power response is stock along with its wavelength. During
the conversion to rgb we normalize these responses, then translate
them to xyz then to rgb.
Two cases need to be covered:

- Illuminant spectrum
- Reflectance spectrum

For illuminant spectrum, we evaluate just the incoming wavelength, and
stock the evaluated value inside the hit record.

For reflectance spectrum, we evaluate the incoming wavelength and
multiply it with the incoming emitted value of the illuminant
spectrum.

The ray color function needs to be overriden for stocking
responses per wavelength. Basically we are going to
add another parameter something like output spectrum.
which stocks power values per wavelength. This might oblige us to
rewrite innerloop as well.