General results
Abstracting away color was a good choice in the namespace but it is not
possible to just use it as a type alias for different sorts of output.
It needs to be implemented as a class. Containg two different data.
Type: Power or Vector.
Data: holds both structures but contains a value in one based on type.
Operators: Some common operators between vec3 and Power.