RGB model works and other roadmap for implementation.
RGB model should be working now.
For sampling spectrum implementation the only thing that
matters is conversions between spectrum values and
rgb values. For that we need to be able to add subtract
spectrums from each other.
The problem with that approach is to have spectrums with
different sizes and wavelengths. Thus we need to be able
to resample spectrums if they have different sizes.
Once that is done the from rgb method of pbr spectrum
can be used and we can output rgb values as well.